pyproject.toml
[project]
name = "pyomechanics"
version = "0.1.0"
description = "Biomechanics in python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ezc3d>=1.5.17",
    "jupyterlab>=4.3.1",
    "kineticstoolkit>=0.15.0",
    "networkx>=3.4.2",
    "numpy>=2.1.3",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["pyomechanics"]  